- Reworking Image Pyramids  * There are two types now.  One is restricted to integer scaling factors and the other allows arbitrary scales  * Internal scales are now declared lazily, simplifying usage  * Started adding updates for the second type - Started adding operations for rescaling an image

+++ feature/src/gecv/alg/tracker/pklt/GenericPkltFeatSelector.java
@@ -72,7 +72,7 @@ public class GenericPkltFeatSelector<InputImage extends ImageBase, DerivativeIma
 	@Override
 	public void compute(List<PyramidKltFeature> active, List<PyramidKltFeature> availableData ) {
 
-		float scaleBottom = image.getScalingAtLayer(0);
+		float scaleBottom = (float)image.getScalingAtLayer(0);
 
 		// exclude active tracks
 		excludeList.reset();
